requirements.txt
Flask==2.3.2
openai==0.27.0
python-dotenv==1.0.0
Flask-Limiter==2.7.0
